f64 -> f32

+++ src/glyph_cache.rs
@@ -77,12 +77,12 @@ impl<'a> GlyphCache<'a> {
         let glyph_size = glyph.advance();
         self.data.get_mut(&size).unwrap().insert(ch, Rc::new(Character {
             offset: [
-                    bitmap_glyph.left() as f64,
-                    bitmap_glyph.top() as f64
+                    bitmap_glyph.left() as f32,
+                    bitmap_glyph.top() as f32
                 ],
             size: [
-                    (glyph_size.x >> 16) as f64,
-                    (glyph_size.y >> 16) as f64
+                    (glyph_size.x >> 16) as f32,
+                    (glyph_size.y >> 16) as f32
                 ],
             texture: texture,
         }));
